+++ scripts/build.js
@@ -93,6 +93,21 @@ const builds = [
       '@type': path.resolve(rootDir, 'src/type'),
     },
   },
+  {
+    name: 'd3',
+    entry: 'src/adapters/d3/index.ts',
+    libName: 'maidrD3',
+    formats: ['es', 'umd'],
+    fileName: format => format === 'es' ? 'd3.mjs' : 'd3.js',
+    emptyOutDir: false,
+    external: [],
+    useReact: false,
+    useDts: true,
+    aliases: {
+      '@adapters': path.resolve(rootDir, 'src/adapters'),
+      '@type': path.resolve(rootDir, 'src/type'),
+    },
+  },
 ];
 
 function createViteConfig(config) {